쿠키 생성, 권한 처리, S3 추가, 그 외 잡다한 것들

+++ src/main/java/com/budong/san/Controller/BuildingForm.java
@@ -10,7 +10,7 @@ public class BuildingForm {
     private String aptDong;
     private String aptHo;
     private String aptType;
-    private int aptSize;
+    private double aptSize;
     private String aptDirection;
     private String aptTransactionType;
     private int aptPrice;
